requirements.txt
tensorflow>=2.10.0
numpy>=1.19.5
matplotlib>=3.5.0
Pillow>=9.0.0
